+++ pybind/minkowski.cpp
@@ -204,12 +204,6 @@ void instantiate(py::module &m) {
 }
 
 void bind_native(py::module &m) {
-#ifndef CPU_ONLY
-  py::class_<GPUMemoryManager>(m, "MemoryManager")
-      .def(py::init<>())
-      .def("resize", &GPUMemoryManager::resize);
-#endif
-
   std::string name = std::string("CoordsKey");
   py::class_<CoordsKey>(m, name.c_str())
       .def(py::init<>())