+++ src/pooling_global_max.cpp
@@ -100,6 +100,8 @@ void GlobalMaxPoolingForwardGPU(at::Tensor in_feat, at::Tensor out_feat,
       in_feat.data<Dtype>(), out_feat.data<Dtype>(), out_nrows,
       num_nonzero.data<int>(), nchannel, get<0>(in_out), get<1>(in_out), d_scr,
       at::cuda::getCurrentCUDAStream());
+
+  p_coords_manager->clearScratchGPUMemory();
 }
 
 template <typename Dtype>