+++ src/pooling_transpose.cpp
@@ -116,7 +116,7 @@ void PoolingTransposeForwardGPU(at::Tensor in_feat, at::Tensor out_feat,
                                 py::object py_out_coords_key,
                                 py::object py_coords_manager) {
   CoordsManager *p_coords_manager = py_coords_manager.cast<CoordsManager *>();
-  const auto& in_out = p_coords_manager->getInOutMapsGPU(
+  const auto &in_out = p_coords_manager->getInOutMapsGPU(
       tensor_strides, strides, kernel_sizes, dilations, region_type, offsets,
       py_in_coords_key, py_out_coords_key, true, true);
 
@@ -126,29 +126,13 @@ void PoolingTransposeForwardGPU(at::Tensor in_feat, at::Tensor out_feat,
   num_nonzero.resize_({out_nrows});
   num_nonzero.zero_();
 
-  // int dtype_mult = dtypeMultiplier<Dtype, int>(), nnz = 0;
-  int nnz = 0;
-  for (const auto &map : get<0>(in_out))
-    nnz += map.size();
-
-  int *d_scr = (int *)p_coords_manager->getScratchGPUMemory(
-      (out_nrows + 1) * sizeof(int) // csr_row
-  );
-
-  Dtype *d_dscr = (Dtype *)p_coords_manager->getScratchGPUMemory2(
-      ((false ? in_feat.size(0) : 0) + // d_ones
-       nnz +                           // d_csr_val
-       in_feat.size(1) * out_nrows     // d_tmp_out_feat
-       ) *
-      sizeof(Dtype));
-
   cusparseHandle_t handle = at::cuda::getCurrentCUDASparseHandle();
   cusparseSetStream(handle, at::cuda::getCurrentCUDAStream());
 
   NonzeroAvgPoolingForwardKernelGPU<Dtype, int>(
       in_feat.data<Dtype>(), in_feat.size(0), out_feat.data<Dtype>(), out_nrows,
       num_nonzero.data<Dtype>(), in_feat.size(1), get<0>(in_out),
-      get<1>(in_out), false, d_scr, d_dscr, handle,
+      get<1>(in_out), false, handle,
       at::cuda::getCurrentCUDAStream());
 }