+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - CoordsMap size initialization updates
 - Added MinkowskiUnion
 - Updated MinkowskiUnion, MinkowskiPruning docs
+- Use cudaMalloc instead of `at::Tensor` for GPU memory management for illegal memory access, invalid arg.
 
 
 ## [0.3.1] - 2019-12-15